+++ python/ray/autoscaler/_private/azure/node_provider.py
@@ -75,8 +75,6 @@ class AzureNodeProvider(NodeProvider):
 
         # cache node objects
         self.cached_nodes = {}
-        self._internal_ip_map = {}
-        self._external_ip_map = {}
 
     @synchronized
     def _get_filtered_nodes(self, tag_filters):
@@ -173,42 +171,6 @@ class AzureNodeProvider(NodeProvider):
         ip = (self._get_cached_node(node_id=node_id)["internal_ip"]
               or self._get_node(node_id=node_id)["internal_ip"])
         return ip
-
-    def get_node_id(self, ip_address, use_internal=False):
-        def find_node_id():
-            if use_internal:
-                return self._internal_ip_map.get(ip_address)
-            else:
-                return self._external_ip_map.get(ip_address)
-
-        with self.lock:
-            if not find_node_id():
-                self._update_ip_maps()
-
-            if not find_node_id():
-                self._update_ip_maps(deep_update=True)
-
-            if not find_node_id():
-                if use_internal:
-                    known_msg = (
-                        f"Known IP addresses: {list(self._internal_ip_map)}")
-                else:
-                    known_msg = (
-                        f"Known IP addresses: {list(self._external_ip_map)}")
-                raise ValueError(f"ip {ip_address} not found. " + known_msg)
-
-        return find_node_id()
-
-    def _update_ip_maps(self, deep_update=False):
-        with self.lock:
-            all_nodes = self.cached_nodes
-
-            if deep_update:
-                all_nodes = self.non_terminated_nodes({})
-
-            for node_id in all_nodes:
-                self._external_ip_map[self.external_ip(node_id)] = node_id
-                self._internal_ip_map[self.internal_ip(node_id)] = node_id
 
     def create_node(self, node_config, tags, count):
         """Creates a number of nodes within the namespace."""
