+++ python/ray/autoscaler/_private/gcp/node_provider.py
@@ -49,8 +49,6 @@ class GCPNodeProvider(NodeProvider):
         # Cache of node objects from the last nodes() call. This avoids
         # excessive DescribeInstances requests.
         self.cached_nodes = {}
-        self._internal_ip_map = {}
-        self._external_ip_map = {}
 
     def non_terminated_nodes(self, tag_filters):
         with self.lock:
@@ -159,42 +157,6 @@ class GCPNodeProvider(NodeProvider):
                 ip = get_internal_ip(node)
 
             return ip
-
-    def get_node_id(self, ip_address, use_internal=False):
-        def find_node_id():
-            if use_internal:
-                return self._internal_ip_map.get(ip_address)
-            else:
-                return self._external_ip_map.get(ip_address)
-
-        with self.lock:
-            if not find_node_id():
-                self._update_ip_maps()
-
-            if not find_node_id():
-                self._update_ip_maps(deep_update=True)
-
-            if not find_node_id():
-                if use_internal:
-                    known_msg = (
-                        f"Known IP addresses: {list(self._internal_ip_map)}")
-                else:
-                    known_msg = (
-                        f"Known IP addresses: {list(self._external_ip_map)}")
-                raise ValueError(f"ip {ip_address} not found. " + known_msg)
-
-        return find_node_id()
-
-    def _update_ip_maps(self, deep_update=False):
-        with self.lock:
-            all_nodes = self.cached_nodes
-
-            if deep_update:
-                all_nodes = self.non_terminated_nodes({})
-
-            for node_id in all_nodes:
-                self._external_ip_map[self.external_ip(node_id)] = node_id
-                self._internal_ip_map[self.internal_ip(node_id)] = node_id
 
     def create_node(self, base_config, tags, count):
         with self.lock:
