+++ python/ray/autoscaler/_private/local/node_provider.py
@@ -156,7 +156,7 @@ class LocalNodeProvider(NodeProvider):
 
     def __init__(self, provider_config, cluster_name):
         NodeProvider.__init__(self, provider_config, cluster_name)
-        self._internal_ip_map = {}
+
         if cluster_name:
             self.state = ClusterState(
                 "/tmp/cluster-{}.lock".format(cluster_name),
@@ -200,30 +200,6 @@ class LocalNodeProvider(NodeProvider):
 
     def internal_ip(self, node_id):
         return socket.gethostbyname(node_id)
-
-    def get_node_id(self, ip_address, use_internal=True):
-        def find_node_id():
-            if use_internal:
-                return self._internal_ip_map.get(ip_address)
-            else:
-                return self._external_ip_map.get(ip_address)
-
-        if not find_node_id():
-            all_nodes = self.non_terminated_nodes({})
-            for node_id in all_nodes:
-                self._external_ip_map[self.external_ip(node_id)] = node_id
-                self._internal_ip_map[self.internal_ip(node_id)] = node_id
-
-        if not find_node_id():
-            if use_internal:
-                known_msg = (
-                    f"Known IP addresses: {list(self._internal_ip_map)}")
-            else:
-                known_msg = (
-                    f"Known IP addresses: {list(self._external_ip_map)}")
-            raise ValueError(f"ip {ip_address} not found. " + known_msg)
-
-        return find_node_id()
 
     def set_node_tags(self, node_id, tags):
         with self.state.file_lock:
