+++ python/ray/autoscaler/node_provider.py
@@ -1,6 +1,5 @@
 import importlib
 import logging
-import json
 import os
 from typing import Any, Dict
 
@@ -185,6 +184,9 @@ class NodeProvider:
     def __init__(self, provider_config, cluster_name):
         self.provider_config = provider_config
         self.cluster_name = cluster_name
+        self._internal_ip_cache = {}
+        self._external_ip_cache = {}
+
 
     def non_terminated_nodes(self, tag_filters):
         """Return a list of node ids filtered by the specified tags dict.
@@ -221,7 +223,33 @@ class NodeProvider:
         raise NotImplementedError
 
     def get_node_id(self, ip_address, use_internal=False):
-        raise NotImplementedError
+        def find_node_id():
+            if use_internal:
+                return self._internal_ip_cache.get(ip_address)
+            else:
+                return self._external_ip_cache.get(ip_address)
+
+        if not find_node_id():
+            self._update_ip_cache()
+
+        if not find_node_id():
+            if use_internal:
+                known_msg = (
+                    f"Known IP addresses: {list(self._internal_ip_cache)}")
+            else:
+                known_msg = (
+                    f"Known IP addresses: {list(self._external_ip_cache)}")
+            raise ValueError(f"ip {ip_address} not found. " + known_msg)
+
+        return find_node_id()
+
+    def _update_ip_cache(self):
+        all_nodes = self.non_terminated_nodes({})
+
+        for node_id in all_nodes:
+            self._external_ip_cache[self.external_ip(node_id)] = node_id
+            self._internal_ip_cache[self.internal_ip(node_id)] = node_id
+
 
     def create_node(self, node_config, tags, count):
         """Creates a number of nodes within the namespace."""
